revamp with spring jdbc

+++ back-end/src/main/java/lk/ijse/dep10/pos/business/custom/impl/ItemBOImpl.java
@@ -8,77 +8,63 @@ import lk.ijse.dep10.pos.dao.custom.ItemDAO;
 import lk.ijse.dep10.pos.dao.custom.OrderDetailDAO;
 import lk.ijse.dep10.pos.dto.ItemDTO;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.util.List;
 import java.util.stream.Collectors;
-
-@Component
+@Service
+@Transactional
 public class ItemBOImpl implements ItemBO {
 
     private final ItemDAO itemDAO;
     private final OrderDetailDAO orderDetailDAO;
     private final Transformer transformer;
-    private final DataSource dataSource;
 
     public ItemBOImpl(ItemDAO itemDAO, OrderDetailDAO orderDetailDAO, Transformer transformer, DataSource dataSource) {
         this.itemDAO = itemDAO;
         this.orderDetailDAO = orderDetailDAO;
         this.transformer = transformer;
-        this.dataSource = dataSource;
     }
 
     @Override
     public void saveItem(ItemDTO itemDTO) throws Exception {
-        try (Connection connection = dataSource.getConnection()) {
-            itemDAO.setConnection(connection);
             if(itemDAO.existsById(itemDTO.getCode())) throw new BusinessException(BusinessExceptionType.DUPLICATE_RECORD,
                     "save failed: Item code: " + itemDTO.getCode() + " already exists");
 
             itemDAO.save(transformer.toItemEntity(itemDTO));
-        }
     }
 
     @Override
     public void updateItem(ItemDTO itemDTO) throws Exception {
-        try (Connection connection = dataSource.getConnection()) {
-            itemDAO.setConnection(connection);
-            itemDAO.update(transformer.toItemEntity(itemDTO));
-        }
+        if (!itemDAO.existsById(itemDTO.getCode()))
+            throw new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                    "Update failed: Item code: " + itemDTO.getCode() + " does not exist");
+        itemDAO.update(transformer.toItemEntity(itemDTO));
     }
 
     @Override
     public void deleteItemByCode(String itemCode) throws Exception {
-        try (Connection connection = dataSource.getConnection()) {
-            itemDAO.setConnection(connection);
-            orderDetailDAO.setConnection(connection);
+        if (orderDetailDAO.existsOrderDetailByItemCode(itemCode))
+            throw new BusinessException(BusinessExceptionType.INTEGRITY_VIOLATION,
+                    "Delete failed: Item code: " + itemCode + " already associated with some orders");
 
-            if(orderDetailDAO.existsOrderDetailByItemCode(itemCode))
-                throw new BusinessException(BusinessExceptionType.INTEGRITY_VIOLATION,
-                        "Delete failed: Item code: " + itemCode + " already associated with some orders");
-            itemDAO.deleteById(itemCode);
-
-        }
+        if (!itemDAO.existsById(itemCode))
+            throw new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                    "Delete failed: Item code: " + itemCode + " does not exist");
+        itemDAO.deleteById(itemCode);
     }
 
     @Override
     public ItemDTO findItemByCode(String itemCode) throws Exception {
-        try (Connection connection = dataSource.getConnection()) {
-            itemDAO.setConnection(connection);
-
-            return itemDAO.findById(itemCode).map(transformer::fromItemEntity)
-                    .orElseThrow(() -> new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
-                            "No Item record found for the code: " + itemCode));
-        }
+        return itemDAO.findById(itemCode).map(transformer::fromItemEntity).orElseThrow(() -> new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                "No item record found for the code: " + itemCode));
     }
 
     @Override
     public List<ItemDTO> findItems(String query) throws Exception {
-        try (Connection connection = dataSource.getConnection()) {
-            itemDAO.setConnection(connection);
-
-            return itemDAO.findAll().stream().map(transformer::fromItemEntity).collect(Collectors.toList());
-        }
+        return itemDAO.findItems(query).stream().map(transformer::fromItemEntity).collect(Collectors.toList());
     }
 }
